+++ src/global.h
@@ -5,7 +5,7 @@
 
 void SysTick_Handler(void);
 void delay(uint32_t ms);
-void setUnixTimestamp(uint8_t year, uint8_t month, uint8_t day, uint8_t hour, uint8_t minute, uint8_t second, uint16_t millisecond);
+void setUnixTimestamp(uint16_t year, uint8_t month, uint8_t day, uint8_t hour, uint8_t minute, uint8_t second, uint16_t millisecond);
 uint64_t getUnixTimestamp(void);
 void incrementUnixTimestamp(uint32_t ms);
